moved version.py to lib/
